모달 open 코드 변경 및 에너지 사용량 요청 컴포넌트 변경
